requirements.txt
apache-airflow==2.9.1
beautifulsoup4==4.12.3
dask==2024.1.0
duckdb==0.10.2
lxml==4.9.3
numpy==1.26.4
pandas==2.2.2
pandera==0.18.0
regex==2023.10.3
selenium==4.20.0
tqdm==4.66.1
undetected_chromedriver==3.5.5
pyarrow
